Fix according to the review of the reviewer
