+++ capitalize_decorator.rb
@@ -1,0 +1,6 @@
+# class CapitalizeDecorator < Decorator
+class CapitalizeDecorator < Decorator
+    def correct_name
+      super.capitalize
+    end
+  end
