+++ decorator.rb
@@ -1,0 +1,11 @@
+# class Decoration < Nameable
+class Decorator < Nameable
+    def initialize(nameable)
+      super
+      @nameable = nameable
+    end
+  
+    def correct_name
+      @nameable.correct_name
+    end
+  end
