+++ person.rb
@@ -1,13 +1,12 @@
 require_relative 'nameable'
-# class Person < Nameable
-class Person
-  def initialize(id, age = 0, name = 'Unknown', parent_permission: true, nameable: nil)
+
+class Person < Nameable
+  def initialize(id, age = 0, name = 'Unknown', parent_permission: true)
     super
     @id = id
     @name = name
     @age = age
     @parent_permission = parent_permission
-    @nameable = nameable
   end
 
   private
@@ -20,6 +19,10 @@ class Person
 
   def can_use_service?
     of_age? || @parent_permission == true
+  end
+
+  def correct_name
+    name
   end
 
   attr_reader :id
