+++ trimmer_decorator.rb
@@ -1,0 +1,10 @@
+# class TrimmerDecorator < Decorator
+class TrimmerDecorator < Decorator
+    def correct_name
+      name = super
+      return name[0, 10] if name.length > 10
+  
+      name
+    end
+  end
+
